float no decimal porque convierte en str

+++ models/price_history_model.py
@@ -1,7 +1,6 @@
 from enum import Enum
 from datetime import datetime
 from pydantic import BaseModel, Field
-from decimal import Decimal
 
 
 class PriceChangeType(str, Enum):
@@ -13,6 +12,6 @@ class PriceHistory(BaseModel):
   id: str = Field(..., alias="_id")
   product_id: str = Field(...)
   type: PriceChangeType = PriceChangeType.PURCHASE
-  price: Decimal = Field(..., decimal_places=2)
+  price: float = Field(...)
   date: datetime = Field(...)
   created_at: datetime = None
