+++ models/purchase_detail_model.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel, Field
 from datetime import datetime
-from decimal import Decimal
 
 
 class Product(BaseModel):
@@ -14,7 +13,7 @@ class PurchaseDetail(BaseModel):
   purchase_id: str = Field(...)
   product: Product = Field(...)
   units: int = Field(...)
-  unity_price: Decimal = Field(..., decimal_places=2)
-  total_price: Decimal = Field(..., decimal_places=2)
+  unity_price: float = Field(...)
+  total_price: float = Field(...)
   created_at: datetime = None
   updated_at: datetime = None
