+++ models/sale_detail_model.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 from pydantic import BaseModel, Field
-from decimal import Decimal
 
 
 class Product(BaseModel):
@@ -15,6 +14,6 @@ class SaleDetail(BaseModel):
   product: Product = Field(...)
   units: int = Field(...)
   unity_price: int = Field(...)
-  total_price: Decimal = Field(..., decimal_places=2)
+  total_price: float = Field(...)
   created_at: datetime = None
   updated_at: datetime = None
